requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
coloredlogs==15.0.1
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
darkdetect==0.8.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.4
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.3
numpy==1.23.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.8.0.87
onnx==1.17.0
onnx-simplifier==0.4.36
onnxruntime==1.21.0
onnxruntime-gpu==1.21.0
onnxslim==0.1.48
openai==1.68.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.0.0
protobuf==6.30.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.19.1
pyparsing==3.2.1
pyphen==0.17.2
PyQt-Fluent-Widgets==1.7.7
PyQt5==5.15.11
PyQt5-Frameless-Window==0.6.0
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.28.1
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sympy==1.13.1
tensorrt==8.6.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.91
ultralytics-thop==2.0.14
urllib3==1.26.13
weasyprint==65.0
webencodings==0.5.1
xcffib==1.7.1
zopfli==0.2.3.post1
